feat: add main page translation
