+++ src/components/EndpointEditor/EndpointEditor.tsx
@@ -1,30 +1,31 @@
+import { useLang } from '@/lib/langContext';
 import { Button } from '../ui/button';
 import { useState } from 'react';
+import { textContent, TextContentType } from '@/lib/langText';
 
 const EndpointEditor = () => {
   const [isEditing, setEditing] = useState(false);
   const [inputValue, setInputValue] = useState(
     'https://rickandmortyapi.com/graphql'
   );
+  const { lang } = useLang();
+  const text = textContent[lang as keyof TextContentType].dashboard;
+
   return (
-    <div className="flex items-center">
+    <div className="flex flex-col items-start">
       {!isEditing ? (
-        <div className="pl-2 mr-5 w-[15vw] text-gray-300 overflow-hidden whitespace-nowrap overflow-ellipsis max-sm:w-full">
+        <div className="pl-2 my-5 w-full h-[40px] text-gray-300 overflow-hidden whitespace-nowrap overflow-ellipsis">
           {inputValue}
         </div>
       ) : (
         <input
-          className="pl-2 mr-5 w-[15vw] h-[40px] bg-white rounded text-black outline-none max-sm:w-full"
+          className="pl-2 my-5 w-full h-[40px] bg-white rounded text-black outline-none"
           value={inputValue}
           onChange={(e) => setInputValue(e.target.value)}
         />
       )}
-      <Button
-        variant="secondary"
-        className="w-[60px] max-sm:w-[80px]"
-        onClick={() => setEditing(!isEditing)}
-      >
-        {isEditing ? 'Save' : 'Edit'}
+      <Button variant="secondary" onClick={() => setEditing(!isEditing)}>
+        {isEditing ? text.url.save : text.url.edit}
       </Button>
     </div>
   );
